Updated dataset.py
